PR comments - renamed Login.v1 to Login - removed Firebase.js under $ROOT/client/ -removed commented code in firebase.js
