+++ client/src/components/Firebase/firebase.js
@@ -1,12 +1,9 @@
 import app from 'firebase/app';
 import "firebase/auth";
-import { functions } from "firebase";
 import * as firebaseConfigLocal from '../../config/firebase.config.local';
 import * as firebaseConfigDev from '../../config/firebase.config.dev';
 import * as firebaseConfigTest from '../../config/firebase.config.test';
 import * as firebaseConfigProd from '../../config/firebase.config.prod';
-
-//require('firebase/auth')
 import "firebase/firestore";
